refactor: rename QuestionField to BuilderField
