+++ src/client/components/builder/CheckboxField.tsx
@@ -13,7 +13,7 @@ import {
 } from '@chakra-ui/react'
 
 import * as checker from '../../../types/checker'
-import { createQuestionField, QuestionFieldComponent } from './QuestionField'
+import { createBuilderField, QuestionFieldComponent } from './BuilderField'
 import { useCheckerContext } from '../../contexts'
 
 import { BuilderActionEnum, ConfigArrayEnum } from '../../../util/enums'
@@ -149,7 +149,7 @@ const PreviewComponent: QuestionFieldComponent = ({ field }) => {
   )
 }
 
-export const CheckboxField = createQuestionField(
+export const CheckboxField = createBuilderField(
   InputComponent,
   PreviewComponent
 )
