+++ src/client/components/builder/NumericField.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { BiHash } from 'react-icons/bi'
 import { useStyles, Box, HStack, VStack, Text, Input } from '@chakra-ui/react'
 
-import { createQuestionField, QuestionFieldComponent } from './QuestionField'
+import { createBuilderField, QuestionFieldComponent } from './BuilderField'
 import { useCheckerContext } from '../../contexts'
 
 import { BuilderActionEnum, ConfigArrayEnum } from '../../../util/enums'
@@ -65,7 +65,4 @@ const PreviewComponent: QuestionFieldComponent = ({ field }) => {
   )
 }
 
-export const NumericField = createQuestionField(
-  InputComponent,
-  PreviewComponent
-)
+export const NumericField = createBuilderField(InputComponent, PreviewComponent)
