+++ src/client/components/builder/index.ts
@@ -1,4 +1,4 @@
-export { createQuestionField } from './QuestionField'
+export { createBuilderField } from './BuilderField'
 export { Navbar } from './Navbar'
 export { FloatingToolbar } from './FloatingToolbar'
 export { NumericField } from './NumericField'
@@ -8,3 +8,4 @@ export { TitleField } from './TitleField'
 export { QuestionsTab } from './QuestionsTab'
 export { LogicTab } from './LogicTab'
 export { PreviewTab } from './PreviewTab'
+export { ActionButton } from './ActionButton'
